updating the dependencies for SDBG 2.2.0_M6 and clean up code
