+++ cloud-cache-local-instance/src/main/java/io/pivotal/data/cloudcachelocalinstance/CloudCacheLocalInstanceApplication.java
@@ -1,0 +1,13 @@
+package io.pivotal.data.cloudcachelocalinstance;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CloudCacheLocalInstanceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CloudCacheLocalInstanceApplication.class, args);
+	}
+
+}
