+++ cloud-cache-local-instance/src/test/java/io/pivotal/data/cloudcachelocalinstance/CloudCacheLocalInstanceApplicationTests.java
@@ -1,0 +1,13 @@
+package io.pivotal.data.cloudcachelocalinstance;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CloudCacheLocalInstanceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
