--- pizza-store-api-completed/gradle.properties
@@ -1,5 +1,0 @@
-log4jSystemRules = 1.17.1
-springBootVersion = 2.1.7.RELEASE
-springBootDataGemFireVersion=1.1.0.RELEASE
-springShellVersion=1.2.0.RELEASE
-lombokVersion=1.18.8
