--- pizza-store-api-completed/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu Feb 01 15:14:54 EST 2018
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.6-all.zip
