--- pizza-store-api-completed/src/main/java/io/pivotal/cloudcache/app/model/Name.java
@@ -1,70 +1,0 @@
-/*
- * Copyright (C) 2018-Present Pivotal Software, Inc. All rights reserved.
- * This program and the accompanying materials are made available under
- * the terms of the under the Apache License, Version 2.0 (the "License”);
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.pivotal.cloudcache.app.model;
-
-import org.springframework.data.gemfire.mapping.annotation.Region;
-import org.springframework.lang.NonNull;
-
-@Region("Name")
-public class Name {
-
-	@NonNull private String name = "";
-
-	  private Name(String name) {
-		    if (name == null) throw new NullPointerException("name can not be null");
-		    this.name = name;
-		  }
-		  	
-	public static Name of(String name) {
-		    return new Name(name);
-	}
-	 
-	public String getName() {
-		return name;
-	}
- 
-	@Override
-	public String toString() {
-		return "Name [name=" + name + "]";
-	}
-
-
-   @Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((name == null) ? 0 : name.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Name other = (Name) obj;
-		if (name == null) {
-			if (other.name != null)
-				return false;
-		} else if (!name.equals(other.name))
-			return false;
-		return true;
-	}
-
-
-}
