--- pizza-store-api-completed/src/main/java/io/pivotal/cloudcache/app/repository/PizzaRepository.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (C) 2018-Present Pivotal Software, Inc. All rights reserved.
- * This program and the accompanying materials are made available under
- * the terms of the under the Apache License, Version 2.0 (the "License”);
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.pivotal.cloudcache.app.repository;
-
-import java.util.List;
-
-import org.springframework.data.gemfire.repository.Query;
-import org.springframework.data.gemfire.repository.query.annotation.Hint;
-import org.springframework.data.gemfire.repository.query.annotation.Limit;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import io.pivotal.cloudcache.app.model.Pizza;
-
-@Repository
-public interface PizzaRepository extends CrudRepository<Pizza, String> {
-	@Limit(10)
-	@Hint("CustomerEmailIndex")
-	@Query("SELECT * FROM /pizza p WHERE p.toppings in customerInfo.email = $1" )
-	List<Pizza> findPizzaOrderByEmailId(String emailId);
-
-}
