+++ pizza-store-api-completed/src/main/java/io/pivotal/data/pizzastoreapi/PizzaStoreApiApplication.java
@@ -1,0 +1,13 @@
+package io.pivotal.data.pizzastoreapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PizzaStoreApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PizzaStoreApiApplication.class, args);
+	}
+
+}
