+++ pizza-store-api-completed/src/main/java/io/pivotal/data/pizzastoreapi/repo/NameRepository.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright (C) 2019-Present Pivotal Software, Inc. All rights reserved.
+ * This program and the accompanying materials are made available under
+ * the terms of the under the Apache License, Version 2.0 (the "License”);
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.pivotal.data.pizzastoreapi.repo;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import io.pivotal.data.pizzastoreapi.model.Name;
+
+@Repository
+public interface NameRepository extends CrudRepository<Name, String> {
+}
