+++ pizza-store-api-completed/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+server.port=8081
+spring.profiles.active=local
